Split buildings_module.py
